Cleaned up ReSharper inspection warnings
* Added the missing streaming-FDR param tags and fixed the RunPercolatorFdr cref
* Removed redundant name qualifiers, casts, and an unused using
* Fixed the pick-candidate StreamWriter using-initializer disposal order

+++ pwiz_tools/Osprey/Osprey.Tasks/MergeNodeTask.cs
@@ -547,7 +547,7 @@ namespace pwiz.Osprey.Tasks
                         || rq < existingB[3]
                         || (rq == existingB[3] && e.Charge < existingB[4]))
                     {
-                        sharedBounds[sk] = new[] { e.ApexRt, e.StartRt, e.EndRt, rq, (double)e.Charge };
+                        sharedBounds[sk] = new[] { e.ApexRt, e.StartRt, e.EndRt, rq, e.Charge };
                     }
                 }
             }
